FROMLIST v1 media: qcom: camss: csiphy-3ph: Make non-Gen2 PHY work again
Commit fbce0ca24c3a ("media: qcom: camss: csiphy-3ph: Move CSIPHY variables
to data field inside csiphy struct") moved a version check that was
previously guarded by a check for Gen2 to csiphy_init without moving the
Gen2 check along with it. Since it was guarded by the Gen2 check, it was
only handling versions that were Gen2 and returned -ENODEV for others since
they shouldn't pass the Gen2 check to begin with.

Check if the CSIPHY is Gen2 before running this version check and assigning
Gen2 lane regs array.

Fixes: fbce0ca24c3a ("media: qcom: camss: csiphy-3ph: Move CSIPHY variables to data field inside csiphy struct")
Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>
Reviewed-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>